factor out react components into their own file
